Refactor cluster
