+++ src/components/molecules/Visualizer/Engine/Cesium/index.tsx
@@ -113,10 +113,7 @@ const Cesium: React.ForwardRefRenderFunction<EngineRef, EngineProps> = (
               ? property.cameraLimiter?.cameraLimitterTargetArea?.height ?? Number.POSITIVE_INFINITY
               : Number.POSITIVE_INFINITY
           }></ScreenSpaceCameraController>
-        <Camera
-          onChange={handleCameraMoveEnd}
-          percentageChanged={property?.cameraLimiter?.cameraLimitterEnabled ? 0.2 : 0.5}
-        />
+        <Camera onChange={handleCameraMoveEnd} percentageChanged={0.2} />
 
         {limiterDimensions && property?.cameraLimiter?.cameraLimitterShowHelper && (
           <Entity>
